Fix on_raw_message_edit

+++ pikalaxbot/cogs/core.py
@@ -131,9 +131,9 @@ class Core(BaseCog):
         await ctx.send(f'Total resources used: {rss:.3f} {unit}')
 
     @commands.Cog.listener()
-    async def on_raw_message_edit(self, payload):
-        channel: discord.TextChannel = self.bot.get_channel(payload.channel_id)
-        if channel.is_news() and channel.permissions_for(channel.guild.me).manage_messages:
+    async def on_raw_message_edit(self, payload: discord.RawMessageUpdateEvent):
+        channel = self.bot.get_channel(payload.channel_id)
+        if channel.guild is not None and channel.is_news() and channel.permissions_for(channel.guild.me).manage_messages:
             message = discord.Message(data=payload.data, state=self.bot._connection, channel=channel)
             if message.content == '[Original Message Deleted]' and message.author.discriminator == '0000':
                 await self.bot._connection.http.delete_message(payload.channel_id, payload.message_id)
